docker-generateconfig/env-requirements.txt
requests==2.32.2
python-dotenv
